+++ src/Infrastructure/Bundle/InfrastructureBundle/Validator/Constraint/Password/ContainsAtLeastOneUppercase.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Validator\Constraint\Password;
+
+use Symfony\Component\Validator\Constraint;
+
+class ContainsAtLeastOneUppercase extends Constraint
+{
+    public $message = 'validators.password.atLeastOneUppercase';
+}