+++ src/Ui/Bundle/MailBundle/Mail/User/UserAccountTemporarilyDisabledMail.php
@@ -1,0 +1,61 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Ui\Bundle\MailBundle\Mail\User;
+
+use Proximum\Vimeet\Application\Components\Mail\UserMail;
+use Proximum\Vimeet\Application\Event\Events;
+use Proximum\Vimeet\Application\View\Participant\ParticipantInfoView;
+use Proximum\Vimeet\Domain\Model\Event;
+
+class UserAccountTemporarilyDisabledMail extends UserMail
+{
+    public const SUBJECT = 'mail.userAccountTemporarilyDisabled.subject';
+    public const TEMPLATE = 'MailBundle:Mail:User/userAccountTemporarilyDisabled.html.twig';
+
+    /** @var string */
+    protected $subject = self::SUBJECT;
+
+    /** @var string */
+    protected $template = self::TEMPLATE;
+
+    /** @var string */
+    protected $messageId = Events::USER_ACCOUNT_TEMPORARILY_DISABLED;
+
+    /** @var bool */
+    protected $sendToEmailTeam = true;
+
+    /**
+     * @param Event               $event
+     * @param string              $sender
+     * @param string              $receiver
+     * @param string              $locale
+     * @param ParticipantInfoView $participantInfoView
+     */
+    public function __construct(
+        Event $event,
+        string $sender,
+        string $receiver,
+        string $locale,
+        ParticipantInfoView $participantInfoView
+    ) {
+        parent::__construct($sender, $receiver, $locale, $event, $participantInfoView);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSubjectParameters()
+    {
+        return [
+            '%event%' => $this->getEvent()->getTitle(),
+        ];
+    }
+}